refactor: 내 프로젝트 조회 API - projects -> users 모듈로 이동
