+++ src/projects/projects.service.ts
@@ -8,7 +8,7 @@ import {
 import { Connection } from 'typeorm';
 import * as _ from 'lodash';
 
-import { OnOffLine, Projects, ProjectsMembers } from '@app/entity';
+import { OnOffLine, Projects } from '@app/entity';
 import { DateFns } from '@app/utils';
 
 import { ProjectsRepository } from './repository';
@@ -24,7 +24,6 @@ import {
   UpdateProjectsParamRequestDto,
   GetProjects,
   GetProjectsResponseDto,
-  GetUserProjects,
 } from './dto';
 import { GetProjectsTechStack } from './type';
 
@@ -33,14 +32,10 @@ export class ProjectsService {
   constructor(
     private readonly connection: Connection,
     private readonly projectsRepository: ProjectsRepository,
-    private readonly projectsMembersRepository: ProjectsMembersRepository,
     private readonly projectsLikeService: ProjectsLikeService,
   ) {
     this.projectsRepository =
       this.connection.getCustomRepository(ProjectsRepository);
-    this.projectsMembersRepository = this.connection.getCustomRepository(
-      ProjectsMembersRepository,
-    );
   }
 
   public async findProjectWithValidate(projectId: number): Promise<Projects> {
@@ -231,32 +226,5 @@ export class ProjectsService {
       projects: _.isEmpty(result[0]) ? null : this.parseProjects(result[0]),
       projectCount: result[1],
     };
-  }
-
-  private parseUserProjects(projects: ProjectsMembers[]): GetUserProjects[] {
-    return projects.map(
-      ({
-        projects: {
-          title: projectTitle,
-          startDate,
-          endDate,
-          userId: projectUserId,
-        },
-        projectId,
-        userId,
-      }: ProjectsMembers): GetUserProjects => ({
-        projectId,
-        projectTitle,
-        startDate,
-        endDate,
-        isProjectOwner: userId === projectUserId,
-      }),
-    );
-  }
-
-  public async getUserProject(userId: number) {
-    const result = await this.projectsMembersRepository.getUserProjects(userId);
-
-    return _.isEmpty(result) ? null : this.parseUserProjects(result);
   }
 }
