+++ src/users/users.controller.ts
@@ -6,8 +6,10 @@ import { UserRequestDto } from '@api/shared/dto/user-request.dto';
 import {
   UsersController as Controller,
   SetUserProfile,
+  GetUserProjects,
 } from './users.controller.decorator';
 import { UsersService } from './users.service';
+import { GetUserProjectsResponseDto } from './dto';
 
 @Controller()
 export class UsersController {
@@ -19,5 +21,14 @@ export class UsersController {
     @User() { userId }: UserRequestDto,
   ) {
     return this.usersService.setUserProfile(imageFile, userId);
+  }
+
+  @GetUserProjects()
+  public async getUserProjects(@User() { userId }: UserRequestDto) {
+    const result = await this.usersService.getUserProjects(userId);
+
+    return new GetUserProjectsResponseDto({
+      userProjects: result,
+    });
   }
 }
