+++ src/users/users.module.ts
@@ -5,9 +5,10 @@ import { UploadImageModule } from '@app/uploadImage';
 import { UsersController } from './users.controller';
 import { UsersRepository } from './repository';
 import { UsersService } from './users.service';
+import { ProjectsMembersModule } from '../projects/members/projects-members.module';
 
 @Module({
-  imports: [UploadImageModule],
+  imports: [UploadImageModule, ProjectsMembersModule],
   controllers: [UsersController],
   providers: [UsersService, UsersRepository],
   exports: [UsersService, UsersRepository],
